Conflicts:
	src/main/java/cn/explink/b2c/vipshop/VipShopGetCwbDataService.java
	src/main/java/cn/explink/service/CwbOrderService.java

+++ src/main/java/cn/explink/service/CwbOrderService.java
@@ -1513,9 +1513,9 @@ public class CwbOrderService extends BaseOrderService {
 	private void validateAppZhongZhuan(String cwb, CwbOrder co, FlowOrderTypeEnum flowOrderTypeEnum) {
 		int changealowflag = this.getChangealowflagByIdAdd(co);
 		if (changealowflag == 1) {
-			CwbApplyZhongZhuan cwbApplyZhongZhuan = this.cwbApplyZhongZhuanDAO.getCwbApplyZhongZhuanByCwb(cwb);
 			//加上非空判断，防止异常产生 --- 刘武强20160612
-			if(cwbApplyZhongZhuan != null){
+			CwbApplyZhongZhuan cwbApplyZhongZhuan = this.cwbApplyZhongZhuanDAO.getCwbApplyZhongZhuanByCwb(cwb);
+			if(cwbApplyZhongZhuan != null){ 
 				if (cwbApplyZhongZhuan.getIshandle() == 0) {
 					throw new CwbException(cwb, flowOrderTypeEnum.getValue(), ExceptionCwbErrorTypeEnum.Weishenhebuxuzhongzhuan, flowOrderTypeEnum.getText());
 				}
@@ -1755,11 +1755,14 @@ public class CwbOrderService extends BaseOrderService {
 			this.validateCwbChongFu(co, scancwb, flowOrderTypeEnum.getValue(), currentbranchid, 0, 0, ExceptionCwbErrorTypeEnum.CHONG_FU_RU_KU);
 		}
 
+		//一票多件是否存在运单号已经为站点到货，但不全部为到货的标志
+		boolean FenZhanDaoHuoFlag = this.getFenZhanDaoHuoFlag(co);
+		
 		boolean newMPSOrder = this.mpsCommonService.isNewMPSOrder(scancwb);
 		// 先更新运单状态，要不然运单流程里面的当前站为0
 		this.mpsOptStateService.updateTransCwbDetailInfo(scancwb, flowOrderTypeEnum, -1L, currentbranchid, 0L);
 
-		if (((co.getCurrentbranchid() == currentbranchid) && (co.getFlowordertype() == flowOrderTypeEnum.getValue())) || (newMPSOrder && (co.getScannum() > 0))) {
+		if (((co.getCurrentbranchid() == currentbranchid) && (co.getFlowordertype() == flowOrderTypeEnum.getValue())) || (newMPSOrder && (co.getScannum() > 0))|| (!newMPSOrder && FenZhanDaoHuoFlag)) {
 			if (co.getScannum() < 1) {
 				this.handleSubstationGoods(user, cwb, scancwb, currentbranchid, requestbatchno, comment, isauto, co, flowOrderTypeEnum, userbranch, isypdjusetranscwb, true, credate, false, false, isAutoSupplyLink);
 			}
@@ -4611,9 +4614,8 @@ public class CwbOrderService extends BaseOrderService {
 			// added shenhongfei 小件员领货扫描 2016-1-12
 			this.orderInterceptService.checkTransCwbIsIntercept(transCwb, FlowOrderTypeEnum.FenZhanLingHuo);
 		}
-		// 是否放行订单号运单号都可以处理		
+		// 是否放行订单号运单号都可以处理
 		this.deliverTakeGoodsMPSReleaseService.validateReleaseCondition(scancwb);
-		
 
 		CwbOrder co = this.cwbDAO.getCwbByCwbLock(cwb);
 
@@ -5434,15 +5436,15 @@ public class CwbOrderService extends BaseOrderService {
 					this.transCwbDao.saveTranscwb(cwb, cwb);
 					this.cwbDAO.saveTranscwbByCwb(transcwb, cwb);
 					this.cwbDAO.saveBackcarnum(transcwbList.size(), cwb);
-					this.cwbDAO.saveSendcarnum(transcwbList.size(), cwb);
+					//this.cwbDAO.saveSendcarnum(transcwbList.size(), cwb);
 				}
 			}else{//当录入或修改快递单号为空时
-				if("".equals(co.getTranscwb())){
+				if(!"".equals(co.getTranscwb()) && nosysyemflag == null){
 					this.transCwbDao.deleteTranscwb(cwb);
 					this.transCwbDao.saveTranscwb(cwb, cwb);
 					this.cwbDAO.saveTranscwbByCwb("", cwb);
 					this.cwbDAO.saveBackcarnum(1, cwb);
-					this.cwbDAO.saveSendcarnum(1, cwb);
+					//this.cwbDAO.saveSendcarnum(1, cwb);
 				}
 			}
 		}
@@ -9204,6 +9206,7 @@ public class CwbOrderService extends BaseOrderService {
 		list = this.cwbDAO.getCwbsBycwbs(inStr.substring(0, inStr.length() - 1) + ")");
 		return list;
 	}
+	
 	public List<CwbOrder> getLabelPrintCwbsByCwbs(List<String> cwbList) {
 		List<CwbOrder> list = new ArrayList<CwbOrder>();
 		StringBuffer inStr = new StringBuffer();
@@ -9666,5 +9669,45 @@ public class CwbOrderService extends BaseOrderService {
 		} else {
 			return null;
 		}
+	}
+	
+	/**
+	 * 
+	 * 根据传入的订单，判断该订单是否是一票多件非集单并且存在一个运单为到货状态且不全为到货
+	 * @author 刘武强
+	 * @param order
+	 * @return
+	 */
+	private boolean getFenZhanDaoHuoFlag(CwbOrder order){
+		if(order == null){//如果order为空，那么直接返回false
+			return false;
+		}
+		boolean mpsSwitch=order.getIsmpsflag()>0;
+		if(mpsSwitch){ //如果为集单模式，直接返回false
+			return false;
+		}
+		String cwb = order.getCwb();
+		long isypdjusetranscwb = this.customerDAO.getCustomerById(order.getCustomerid()).getCustomerid() == 0 ? 0 : this.customerDAO.getCustomerById(order.getCustomerid()).getIsypdjusetranscwb();
+		if (((order.getSendcarnum() > 1) || (order.getBackcarnum() > 1)) && ((order.getTranscwb().split(",").length > 1) || (order.getTranscwb().split(":").length > 1)) && (isypdjusetranscwb == 1)) {
+			if (order.getTranscwb().indexOf(",") > -1) {
+				String[] transCwbs = order.getTranscwb().split(",");
+				int i = transCwbs.length;
+				for (String transcwb : transCwbs) {//遍历所有的运单，查看是否有运单的状态为到货
+					if ("".equals(transcwb.trim())) {
+						continue;
+					}
+					//从运单轨迹表中得到运单当前的状态记录
+					TranscwbOrderFlow tcof = this.transcwborderFlowDAO.getTranscwbOrderFlowByCwbAndState(transcwb, cwb);
+					
+					if(tcof.getFloworderid() == FlowOrderTypeEnum.FenZhanDaoHuoSaoMiao.getValue()){
+						i -- ;
+					}
+				}
+				if(i != 0 && i < transCwbs.length){//如果订单不全为到货，只是有部分为到货，那么返回true
+					return true;
+				}
+			}
+		}		
+		return false;
 	}
 }
